10270: parameterize smoketest s3 bucket, add cypress env variables
